Split up webpack files for better readability
