+++ package.json
@@ -24,13 +24,13 @@
 	},
 	"license": "GPL-3.0+",
 	"scripts": {
-		"build": "cross-env BABEL_ENV=default NODE_ENV=production webpack",
+		"build": "rimraf build/* && cross-env BABEL_ENV=default NODE_ENV=production webpack",
 		"build:e2e-test": "npm run build",
 		"build:map": "cross-env BABEL_ENV=default NODE_ENV=production FORCE_MAP=true webpack",
 		"changelog": "node ./bin/changelog",
 		"changelog:zenhub": "node ./bin/changelog --changelogSrcType='ZENHUB_RELEASE'",
 		"deploy": "cross-env WOOCOMMERCE_BLOCKS_PHASE=2 composer install --no-dev && npm run build --loglevel error && sh ./bin/github-deploy.sh",
-		"dev": "cross-env BABEL_ENV=default webpack",
+		"dev": "rimraf build/* && cross-env BABEL_ENV=default webpack",
 		"explore": "source-map-explorer",
 		"lint": "npm run lint:php && npm run lint:css && npm run lint:js",
 		"lint:ci": "npm run lint:js && npm run lint:css",
@@ -110,7 +110,6 @@
 		"babel-plugin-transform-react-remove-prop-types": "0.4.24",
 		"babel-plugin-transform-runtime": "6.23.0",
 		"chalk": "3.0.0",
-		"clean-webpack-plugin": "3.0.0",
 		"commander": "4.1.1",
 		"create-file-webpack": "1.0.2",
 		"cross-env": "6.0.3",
@@ -140,7 +139,7 @@
 		"puppeteer": "2.1.1",
 		"react-test-renderer": "16.13.0",
 		"request-promise": "4.2.5",
-		"rimraf": "3.0.2",
+		"rimraf": "^3.0.2",
 		"sass-loader": "7.3.1",
 		"source-map-explorer": "2.4.2",
 		"stylelint": "12.0.1",
